+++ src/components/studio/index.ts
@@ -6,6 +6,7 @@ export { default as ProjectActions } from "./ProjectActions";
 export { default as ProjectProgress } from "./ProjectProgress";
 export { default as ProjectStatusCards } from "./ProjectStatusCards";
 export { default as PipelineStatus } from "./PipelineStatus";
+export { default as VideoPanel } from "./VideoPanel";
 export { default as AudioPanel } from "./AudioPanel";
 export { default as ThumbnailPanel } from "./ThumbnailPanel";
 export { default as SEOPanel } from "./SEOPanel";